+++ src/Java/myservicebus-testing/src/main/java/com/myservicebus/TestingServiceExtensions.java
@@ -21,4 +21,3 @@ public class TestingServiceExtensions {
         return services;
     }
 }
-